fix some nasty bugs
